reserves opt. on L1

+++ src/L1/interfaces/IDove.sol
@@ -10,7 +10,7 @@ interface IDove {
     event Claim(address indexed recipient, uint256 amount0, uint256 amount1);
     event FeesUpdated(address recipient, uint256 amount0, uint256 amount1);
     event FeesTransferred(address indexed sender, uint256 amount0, uint256 amount1, address indexed to);
-    event Updated(uint256 reserve0, uint256 reserve1);
+    event Updated(uint128 reserve0, uint128 reserve1);
     event Bridged(uint256 indexed srcChainId, uint256 syncId, address token, uint256 amount);
     event SyncPending(uint256 indexed srcDomain, uint256 syncID);
     event SyncFinalized(
@@ -50,5 +50,5 @@ interface IDove {
     function syncL2(uint32 destinationDomain, address pair) external payable;
     function finalizeSyncFromL2(uint32 originDomain, uint256 syncID) external;
     function claimBurn(uint32 srcDomain, address user) external;
-    function getReserves() external view returns (uint256 reserve0, uint256 reserve1);
+    function getReserves() external view returns (uint128 reserve0, uint128 reserve1);
 }
